Added support for random UID generator based on recommendations in RFC7986
