+++ src/test/groovy/net/fortuna/ical4j/util/FixedUidGeneratorSpec.groovy
@@ -2,14 +2,14 @@ package net.fortuna.ical4j.util
 
 import spock.lang.Specification
 
-class UidGeneratorSpec extends Specification {
+class FixedUidGeneratorSpec extends Specification {
 
     def "verify host info is cached"() {
         given: 'a mock host info'
         def hostInfo = Mock(HostInfo)
 
         when: 'a generator is created and a uid is generated multiple times'
-        UidGenerator generator = [hostInfo, '1']
+        FixedUidGenerator generator = [hostInfo, '1']
         1..10.each { generator.generateUid() }
 
         then: 'HostInfo.getHostName() is only called once'
