+++ src/test/java/net/fortuna/ical4j/model/component/AvailableTest.java
@@ -31,8 +31,6 @@
  */
 package net.fortuna.ical4j.model.component;
 
-import java.net.SocketException;
-
 import junit.framework.TestSuite;
 import net.fortuna.ical4j.model.ComponentTest;
 import net.fortuna.ical4j.model.DateTime;
@@ -40,7 +38,10 @@ import net.fortuna.ical4j.model.Dur;
 import net.fortuna.ical4j.model.property.DtStamp;
 import net.fortuna.ical4j.model.property.DtStart;
 import net.fortuna.ical4j.model.property.Duration;
+import net.fortuna.ical4j.util.RandomUidGenerator;
 import net.fortuna.ical4j.util.UidGenerator;
+
+import java.net.SocketException;
 
 /**
  * $Id$
@@ -69,7 +70,7 @@ public class AvailableTest extends ComponentTest {
         suite.addTest(new AvailableTest("testIsNotCalendarComponent", a));
         suite.addTest(new AvailableTest("testValidationException", a));
         
-        UidGenerator g = new UidGenerator("test");
+        UidGenerator g = new RandomUidGenerator();
         a = new Available();
         a.getProperties().add(g.generateUid());
         a.getProperties().add(new DtStart(new DateTime()));
